refactor: 홈 프로젝트를 MDX 단일 소스 기반으로 전환 (page)
홈 page.tsx를 서버 컴포넌트로 전환하고 본문은 home-view로 분리.
제목·태그·아이콘은 _projects MDX에서 가져오고, 홈 전용 순서·단축 설명만
featured-projects.ts에서 관리하여 데이터 중복 제거
