+++ app/page.tsx
@@ -1,324 +1,27 @@
-"use client";
+import { getAllProjects } from "@/_projects";
+import { HomeView } from "@/app/home-view";
+import { featuredProjects, type HomeProject } from "@/lib/featured-projects";
 
-import { PageContainer } from "@/components/page-container";
-import { SkillBadge } from "@/components/skill-badge";
-import { TextType } from "@/components/text-type";
-import { sectionReveal, staggerContainer, staggerItem } from "@/lib/animation";
-import { skillCategories } from "@/lib/skill-data";
-import { timeline } from "@/lib/timeline-data";
-import { cn } from "@/lib/utils";
-import { ArrowUpRight, Code } from "lucide-react";
-import { motion } from "motion/react";
-import Image from "next/image";
-import Link from "next/link";
+const HomePage = async () => {
+  const all = await getAllProjects();
+  const bySlug = new Map(all.map((p) => [p.slug, p]));
 
-// About Data
-const aboutPhrases = [
-  "빠르게 습득하고, 확실하게 쓰는 개발자 안성진입니다.",
-  "교육·실무 어디서든 상급 이상을 달성해온 개발자 안성진입니다.",
-  "조기 출근, 야근, 긴급 소집에 익숙한 개발자 안성진입니다.😉",
-  "先入後出(선입후출), STACK처럼 가장 처음부터 끝까지 마무리하는 개발자 안성진입니다.",
-  "Java에서 시작해 Next.js·TypeScript·Prisma로 전환한 개발자 안성진입니다.",
-  "Vanilla JS만으로도 웹을 만들 수 있는 개발자 안성진입니다.",
-  "PHP를 React, Remix.js로 마이그레이션한 개발자 안성진입니다.",
-];
+  // _projects(MDX)를 단일 소스로, featured-projects의 순서·단축 설명만 덧입힌다
+  const projects: HomeProject[] = featuredProjects
+    .map((f) => {
+      const p = bySlug.get(f.slug);
+      if (!p) return null;
+      return {
+        slug: f.slug,
+        title: p.frontmatter.title as string,
+        summary: f.summary,
+        tags: (p.frontmatter.tags as string[]) ?? [],
+        icon: p.icon,
+      };
+    })
+    .filter((p): p is HomeProject => p !== null);
 
-// Projects Data
-const projects = [
-  {
-    title: "육아밸 (yougabell)",
-    description: "워킹맘/워킹대디를 위한 육아 정보·기록·AI 챗봇 서비스",
-    tags: ["NestJS", "Next.js", "Expo", "Supabase"],
-    slug: "yougabell",
-    icon: "/images/projects/yougabell/icon.png",
-  },
-  {
-    title: "Reciflow",
-    description: "레시피 공유·관리 모바일 플랫폼 (Supabase 백엔드 중심)",
-    tags: ["Supabase", "PostgreSQL", "Expo", "Next.js"],
-    slug: "reciflow",
-    icon: "/images/projects/reciflow/icon.png",
-  },
-  {
-    title: "딸깍톤 (Ttalkkakthon)",
-    description: "AI와 함께하는 해커톤 — 참가 신청·갤러리·디스코드 기반 AI 자동 심사를 지원하는 해커톤 운영 플랫폼",
-    tags: ["Next.js", "TypeScript", "Prisma", "Neon"],
-    slug: "ttalkkakthon",
-    icon: "/images/projects/ttalkkakthon/icon.svg",
-  },
-  {
-    title: "DPS (Design Plug Shop)",
-    description: "판매자와 제조사를 연결하는 B2B 주문형 굿즈·인쇄 제작 플랫폼",
-    tags: ["Next.js", "TypeScript", "Prisma", "NextAuth"],
-    slug: "dps",
-    icon: "/images/projects/dps/icon.webp",
-  },
-  {
-    title: "DPS Store (디플샵 스토어)",
-    description: "멀티테넌트 기반 팝업 스토어 플랫폼 — 테넌트별 독립 온라인 스토어 운영",
-    tags: ["Next.js", "TypeScript", "Prisma", "Caddy"],
-    slug: "dps-store",
-    icon: "/images/projects/dps-store/icon.svg",
-  },
-  {
-    title: "정보보안기사 시험 대비 웹 앱",
-    description: "이론 학습, 문제 풀이, 모의고사 기능을 갖춘 PWA 웹 애플리케이션",
-    tags: ["React", "TypeScript", "Vite", "PWA"],
-    slug: "information-security-engineer",
-    icon: "/images/projects/information-security-engineer/icon.svg",
-  },
-  {
-    title: "NAYA",
-    description: "멀티미디어 소개 카드와 명함을 제작·공유·관리하는 크로스 플랫폼 서비스",
-    tags: ["Kotlin", "Jetpack Compose", "CameraX", "Room", "React", "Spring Boot"],
-    slug: "naya",
-    icon: "/images/projects/naya/icon.svg",
-  },
-  {
-    title: "README",
-    description: "그림 퀴즈 게임으로 NFT를 생성하고 거래하는 블록체인 기반 플랫폼",
-    tags: ["React", "TypeScript", "Redux", "Socket.io", "Web3.js", "Solidity"],
-    slug: "readme-nft",
-    icon: "/images/projects/readme-nft/icon.svg",
-  },
-  {
-    title: "당신의 계절",
-    description: "퍼스널 컬러 자가 진단 및 전문 컨설턴트 1:1 화상 진단 서비스",
-    tags: ["React", "Redux", "OpenVidu", "Spring Boot", "Redis"],
-    slug: "your-seasons",
-    icon: "/images/projects/your-seasons/icon.webp",
-  },
-  {
-    title: "SSAFIT",
-    description: "운동 영상 관리, 운동 기록 추적, 식단 관리를 통합한 피트니스 웹 애플리케이션",
-    tags: ["Vue.js", "Vuex", "Chart.js", "Spring Boot", "MyBatis", "MySQL"],
-    slug: "ssafit",
-    icon: "/images/projects/ssafit/icon.svg",
-  },
-];
-
-// Contact Data
-const contactLinks = [
-  {
-    title: "Email",
-    href: "mailto:hello@anveloper.dev",
-    value: "hello@anveloper.dev",
-  },
-  {
-    title: "GitHub",
-    href: "https://github.com/anveloper",
-    value: "anveloper",
-  },
-  // TODO: 수정중
-  // {
-  //   title: "Resume",
-  //   href: "https://anveloper-dev.notion.site/ca13ffc984be4ce399d73659aebbe303",
-  //   value: "Notion",
-  // },
-];
-
-const SectionHeader = ({ children, href }: { children: React.ReactNode; href?: string }) => {
-  const heading = (
-    <h2 className="text-[11px] md:text-xs font-bold uppercase tracking-[0.2em] text-muted-foreground/60">{children}</h2>
-  );
-  if (href) {
-    return (
-      <Link href={href} className="group flex items-center gap-1.5 mb-4 md:mb-6 w-fit">
-        {heading}
-        <ArrowUpRight className="w-3 h-3 text-muted-foreground opacity-0 -translate-y-0.5 -translate-x-0.5 transition-all group-hover:opacity-100 group-hover:translate-y-0 group-hover:translate-x-0" />
-      </Link>
-    );
-  }
-  return <div className="mb-4 md:mb-6">{heading}</div>;
-};
-
-const viewportConfig = { once: true, margin: "-80px" as const };
-
-const HomePage = () => {
-  return (
-    <PageContainer>
-      {/* Hero Section */}
-      <motion.section variants={sectionReveal} initial="hidden" animate="visible" className="mb-10 md:mb-16">
-        <div className="flex items-center gap-4 md:gap-6">
-          <div className="relative shrink-0">
-            <div className="relative w-14 h-14 md:w-16 md:h-16 rounded-full overflow-hidden ring-2 ring-primary-sky/20">
-              <Image src="/profile.webp" alt="안성진 프로필" fill className="object-cover" priority />
-            </div>
-            <div className="absolute -bottom-1 -right-1 w-4 h-4 md:w-5 md:h-5 bg-primary-sky rounded-full border-2 border-background flex items-center justify-center">
-              <Code className="w-2 h-2 md:w-2.5 md:h-2.5 text-white" />
-            </div>
-          </div>
-          <div className="flex flex-col gap-0.5 md:gap-1">
-            <h1 className="text-2xl md:text-3xl font-bold text-foreground tracking-tight">안성진</h1>
-            <p className="text-xs md:text-sm font-semibold text-muted-foreground">
-              Full-Stack Developer, UX Team Lead / Manager
-            </p>
-            <p className="text-[11px] md:text-xs text-muted-foreground/80 italic tracking-tight">
-              언어로 세상을 표현하는 개발자
-            </p>
-          </div>
-        </div>
-      </motion.section>
-
-      {/* About Section */}
-      <motion.section
-        variants={sectionReveal}
-        initial="hidden"
-        whileInView="visible"
-        viewport={viewportConfig}
-        className="mb-10 md:mb-16"
-      >
-        <p className="text-sm md:text-base leading-relaxed text-foreground min-h-[7rem] sm:min-h-[5.5rem] md:min-h-[5rem]">
-          오늘의 쓸모를 다하는 사람,
-          <br />
-          <TextType
-            text={aboutPhrases}
-            typingSpeed={40}
-            deletingSpeed={25}
-            pauseDuration={2000}
-            className="font-semibold"
-            highlightText="개발자 안성진"
-            highlightClassName="text-primary-sky"
-            cursorClassName="text-primary-sky"
-          />
-        </p>
-      </motion.section>
-
-      {/* Skills Section */}
-      <motion.section
-        variants={sectionReveal}
-        initial="hidden"
-        whileInView="visible"
-        viewport={viewportConfig}
-        className="mb-10 md:mb-16"
-      >
-        <SectionHeader href="/skills">Technical Stack</SectionHeader>
-        <div className="flex flex-col gap-3 md:gap-4">
-          {skillCategories.map((category) => (
-            <div key={category.title} className="flex items-start gap-3 md:gap-4">
-              <span className="w-20 md:w-28 pt-1 text-[11px] md:text-xs font-bold text-primary-sky shrink-0 uppercase">
-                {category.title}
-              </span>
-              <div className="flex flex-wrap gap-1.5 md:gap-2">
-                {category.skills.map((skill) => (
-                  <SkillBadge key={skill.name} skill={skill} />
-                ))}
-              </div>
-            </div>
-          ))}
-        </div>
-      </motion.section>
-
-      {/* Experience & Education Section — Compact Timeline */}
-      <motion.section
-        variants={sectionReveal}
-        initial="hidden"
-        whileInView="visible"
-        viewport={viewportConfig}
-        className="mb-10 md:mb-16"
-      >
-        <SectionHeader href="/education">Milestones</SectionHeader>
-        <div className="flex flex-col border-l border-border/30 ml-1 pl-4 md:pl-6 gap-4 md:gap-5">
-          {timeline.map((item, index) => (
-            <div key={`${item.title}-${index}`} className="relative">
-              <div
-                className={cn(
-                  "absolute -left-5.25 md:-left-6.75 top-1 w-2.5 h-2.5 md:w-3 md:h-3 rounded-full ring-4 ring-background",
-                  index === 0 ? "bg-primary-sky" : "bg-muted-foreground/30"
-                )}
-              />
-              <p className="text-xs md:text-sm font-bold text-foreground leading-snug">
-                {item.title} · {item.organization}
-              </p>
-              <p
-                className={cn(
-                  "text-[11px] md:text-xs mt-1",
-                  index === 0 ? "text-primary-sky" : "text-muted-foreground"
-                )}
-              >
-                {item.period}
-              </p>
-            </div>
-          ))}
-        </div>
-      </motion.section>
-
-      {/* Projects Section */}
-      <motion.section
-        variants={sectionReveal}
-        initial="hidden"
-        whileInView="visible"
-        viewport={viewportConfig}
-        className="mb-10 md:mb-16"
-      >
-        <SectionHeader href="/projects">Featured Projects</SectionHeader>
-        <motion.div
-          variants={staggerContainer}
-          initial="hidden"
-          whileInView="visible"
-          viewport={viewportConfig}
-          className="grid grid-cols-1 sm:grid-cols-2 gap-2 md:gap-3"
-        >
-          {projects.map((project, index) => (
-            <motion.div
-              key={project.slug}
-              variants={staggerItem}
-              className={index === projects.length - 1 && projects.length % 2 !== 0 ? "col-span-2" : ""}
-            >
-              <Link
-                href={`/projects/${project.slug}`}
-                className={cn(
-                  "block group p-3 md:p-4 bg-secondary/50 rounded-lg hover:bg-accent transition-colors",
-                  "flex flex-col gap-2",
-                  index === projects.length - 1 && projects.length % 2 !== 0 && "flex-row items-center gap-4"
-                )}
-              >
-                {project.icon && (
-                  <img src={project.icon} alt="" className="w-8 h-8 md:w-9 md:h-9 rounded-md shrink-0" />
-                )}
-                <div className="flex-1 min-w-0">
-                  <h3 className="text-sm md:text-base font-bold text-foreground group-hover:text-primary-sky transition-colors truncate">
-                    {project.title}
-                  </h3>
-                  <p className="text-xs md:text-sm text-muted-foreground leading-snug mt-0.5 line-clamp-1">
-                    {project.description}
-                  </p>
-                </div>
-                <div className="flex flex-wrap gap-1">
-                  {project.tags.slice(0, 3).map((tag) => (
-                    <span
-                      key={tag}
-                      className="text-[10px] md:text-xs px-1.5 py-0.5 bg-background text-muted-foreground rounded-md"
-                    >
-                      {tag}
-                    </span>
-                  ))}
-                </div>
-              </Link>
-            </motion.div>
-          ))}
-        </motion.div>
-      </motion.section>
-
-      {/* Contact Section */}
-      <motion.section variants={sectionReveal} initial="hidden" whileInView="visible" viewport={viewportConfig}>
-        <SectionHeader>Connect</SectionHeader>
-        <div className="flex flex-col gap-2">
-          {contactLinks.map((link) => (
-            <a
-              key={link.title}
-              href={link.href}
-              {...(link.title !== "Email" ? { target: "_blank", rel: "noopener noreferrer" } : {})}
-              className="flex items-center gap-3 p-3 bg-secondary rounded-lg hover:bg-accent transition-colors"
-            >
-              <span className="text-[11px] md:text-sm font-medium text-foreground">{link.value}</span>
-              <ArrowUpRight className="w-3 h-3 text-muted-foreground/40 ml-auto" />
-            </a>
-          ))}
-        </div>
-      </motion.section>
-    </PageContainer>
-  );
+  return <HomeView featuredProjects={projects} />;
 };
 
 export default HomePage;
